fix(ui,shader): light-theme paint on auth/settings/composer + SKSL casts
Phase 9 Block F continuation — light theme finishes on logic-heavy
files where Phase 8 had kept business logic intact.

app/auth/index.tsx (97 line changes):
- Form input borders: hairline rgba(15,22,36,0.10), focused #2DD4BF
- Input backgrounds: white with hairline (or backgroundSubtle for filled)
- Email Continue button: teal #2DD4BF bg, white text, 54px height
- Eyebrow text: mono uppercase letterspaced, secondary color
- Error text: depleted #DC2626
- Apple Sign In button kept solid black per HIG (untouched)
- Preserved: AppleAuthenticationButton swap, textContentType + autoComplete
  autofill, refs for next-key-focus, Haptics on success

app/settings.tsx (128 line changes):
- Header logo container: white bg + hairline border, dynamic chip via
  brandAccent (already wired)
- Demo + Free-User-View banners: amber/translucent with primary text
- Section labels: mono uppercase letterspaced, textSecondary
- SettingsRow: hairline divider, primary text, ChevronRight in tertiary
- Modal pickers (language, demo): white bg, hairline border + dividers
- Logout: text-only depleted #DC2626
- Preserved: useGlassStyle hook, useBrandAccent, Phase 5 a11y wrapping,
  Phase 6 sub-brand wiring, Platform Beta entries, all callbacks +
  destructive haptic

components/Composer.tsx (4 line changes):
- Submit button bg: effectiveAccent solid (was 20% alpha mix) when active,
  colors.backgroundSubtle when idle
- Border: effectiveAccent solid or colors.hairline
- Preserved: useGlassStyle, useBrandAccent, all submit logic

components/atmosphericShader.ts (6 line changes):
- SKSL type correctness — mix() second arg requires explicit half() cast
  in spectrum() helper. Self-correction from Block A agent post-write.
- No behavior change, prevents shader compile error on Skia runtime.

All Pressables a11y preserved. Capacity-spectrum hex preserved verbatim.
StatusBar style="dark" content where light surfaces apply.

+++ components/atmosphericShader.ts
@@ -275,12 +275,12 @@ SceneSample mapScene(float3 p, float t) {
 half3 spectrum(float t, float hueShift) {
     float u = fract(t + hueShift / 6.2831);
     if (u < 0.3333) {
-        return mix(CRIMSON, AMBER, u / 0.3333);
+        return mix(CRIMSON, AMBER, half(u / 0.3333));
     }
     if (u < 0.6666) {
-        return mix(AMBER, TEAL, (u - 0.3333) / 0.3333);
+        return mix(AMBER, TEAL, half((u - 0.3333) / 0.3333));
     }
-    return mix(TEAL, CYAN, (u - 0.6666) / 0.3334);
+    return mix(TEAL, CYAN, half((u - 0.6666) / 0.3334));
 }
 
 //
